Speed up and allow for different pin assignments
This mostly replaces the _send() code with a speed optimised variant. A major
improvement comes from avoiding unnecessary delays and  beginTransmission() / endTransmission() pairs.
In addition, we add an additional write() method with further optimisations
for text strings. Refreshing the complete display is now faster by a factor
of about three.

While at it we add new constructors and class variables to allow changing the
pin layout.  Known modules are for now the default and an Raspberry Pi Module by
IN-JOY, which works with the Arduino as well.